Online Voting System
Using Online Voting System, you can hold an election for various
positions at  your organisation and then the nominees can register
themselves using the OrganisationNumber and can nominate  for these
position. The voter can register themselves using the OrganisationNumber
and  can vote to candidates  standing in the election of their
organization. The result of the election can be seen during as well as
after the election. After login one see the schedule, nominees and
result  of election of their organisation.